backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.8
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.2
django-cors-headers==4.7.0
django-formtools==2.5.1
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-tenants==3.8.0
django-two-factor-auth==1.17.0
djangorestframework==3.16.0
frozenlist==1.6.2
idna==3.10
multidict==6.4.4
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
pypng==0.20220715.0
qrcode==7.4.2
requests==2.32.3
sqlparse==0.5.3
twilio==9.6.2
typing_extensions==4.14.0
urllib3==2.4.0
yarl==1.20.0
